requirements.txt
fastapi
uvicorn
pandas
matplotlib
sentence-transformers
faiss-cpu
transformers
torch
huggingface_hub